long id

+++ src/main/java/web/controller/UserController.java
@@ -39,7 +39,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public String showUser(@PathVariable ("id") int id, Model model){
+    public String showUser(@PathVariable ("id") long id, Model model){
         model.addAttribute("user", userService.getUserById(id));
         return "user";
     }
@@ -51,7 +51,7 @@ public class UserController {
     }
 
     @RequestMapping("/users/remove/{id}")
-    public String removeUser(@PathVariable("id") int id){
+    public String removeUser(@PathVariable("id") long id){
         userService.removeUser(id);
         return "redirect:/users";
     }
